# speedup tests: no full database migrations in setup