--- a/tests/Propel/Tests/Generator/Command/util/migrate-to-version/version-2/schema.xml
+++ b/tests/Propel/Tests/Generator/Command/util/migrate-to-version/version-2/schema.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="ISO-8859-1" standalone="no"?>
-<database name="migration_command">
+<database name="migration_command" schema="migration_command_db">
 
     <table name="table1">
         <column name="id" required="true" primaryKey="true" autoIncrement="true" type="INTEGER" description="Book Id"/>