Running database tests locally takes a long time. Turns out the tests for migration command work on the bookstore database, but five of them start by migrating the bookstore schema to a two-table schema, and then end with restoring the full bookstore schema.

This takes up a long time, which is annoying during development, and it doesn't even help, the extensive migration is not part of the tests. You can, however, screw up your database when aborting running tests, because when aborting during those tests (which is likely, considering how long they run) the bookstore database won't be migrated back.

On my laptop, the five tests each take around 15s, almost 1.5 minutes in total:
```xml
<?xml version="1.0" encoding="UTF-8"?>
<testsuites>
  <testsuite name="" tests="1322" assertions="4227" errors="0" warnings="5" failures="0" skipped="17" time="138.882636">
    <testsuite name="propel2" tests="1322" assertions="4227" errors="0" warnings="5" failures="0" skipped="17" time="138.882636">
      ...
      <testsuite name="Propel\Tests\Generator\Command\MigrationTest" file="./tests/Propel/Tests/Generator/Command/MigrationTest.php" tests="14" assertions="35" errors="0" warnings="0" failures="0" skipped="0" time="72.571424">
        ...
        <testcase name="testMigrateCommandShouldDoNothingIfGivenVersionIsTheLastAppliedVersion"... time="12.718369"/>
        <testcase name="testMigrateCommandShouldRollbackToTheGivenVersionIfItIsLowerThanTheCurrentVersion"... time="14.391237"/>
        <testcase name="testMigrateCommandShouldMigrateToTheGivenVersionIfItIsHigherThanTheCurrentVersion"... time="14.552978"/>
        <testcase name="testMigrationStatusCommandShouldReturnEmptyWhenOptionIsProvidedAndMigrationsWereNotExecuted"... time="0.038497"/>
        <testcase name="testMigrationStatusCommandShouldReturnTheLastMigrationVersionWhenOptionIsProvided"... time="14.465734"/>
        <testcase name="testMigrationStatusCommandShouldNotReturnTheLastMigrationVersionWhenOptionIsNotProvided"... time="14.669982"/>
      </testsuite>
```
Having them use an empty database (this PR), it goes down significantly, now each test is only a bit over 2s:
```xml
<?xml version="1.0" encoding="UTF-8"?>
<testsuites>
  <testsuite name="" tests="1322" assertions="4227" errors="0" warnings="5" failures="0" skipped="17" time="68.031552">
    <testsuite name="propel2" tests="1322" assertions="4227" errors="0" warnings="5" failures="0" skipped="17" time="68.031552">
      ...
      <testsuite name="Propel\Tests\Generator\Command\MigrationTest" file="./tests/Propel/Tests/Generator/Command/MigrationTest.php" tests="14" assertions="35" errors="0" warnings="0" failures="0" skipped="0" time="10.755472">
        ...
        <testcase name="testMigrateCommandShouldDoNothingIfGivenVersionIsTheLastAppliedVersion"... time="2.003465"/>
        <testcase name="testMigrateCommandShouldRollbackToTheGivenVersionIfItIsLowerThanTheCurrentVersion"... time="2.116104"/>
        <testcase name="testMigrateCommandShouldMigrateToTheGivenVersionIfItIsHigherThanTheCurrentVersion"... time="2.113315"/>
        <testcase name="testMigrationStatusCommandShouldReturnEmptyWhenOptionIsProvidedAndMigrationsWereNotExecuted"... time="0.026195"/>
        <testcase name="testMigrationStatusCommandShouldReturnTheLastMigrationVersionWhenOptionIsProvided"... time="2.043945"/>
        <testcase name="testMigrationStatusCommandShouldNotReturnTheLastMigrationVersionWhenOptionIsNotProvided"... time="2.223026"/>
      </testsuite>
```
This halves total execution time (from 140s to 70s).

Weirdly enough, it does not affect runtime on GitHub, not sure why. Tests take a good minute, same as before. My guess is that they were never affected by the issue, either because tests are executed in a different order, or because it just doesn't matter when there is no disk access involved.